pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "typeslate-terminal"
version = "1.0.0"
description = "TypeSlate Terminal - A distraction-free writing environment for the terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["writing", "terminal", "tui", "distraction-free", "text-editor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Editors",
]
dependencies = [
    "textual>=0.85.0",
    "pyspellchecker>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/josh-writes/typeslate-terminal"
Repository = "https://github.com/josh-writes/typeslate-terminal"
Issues = "https://github.com/josh-writes/typeslate-terminal/issues"

[project.scripts]
typeslate = "typeslate_terminal.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
